pyproject.toml
[tool.poetry]
name = "twitter-back"
version = "0.1.0"
description = "Twitter Api"
authors = ["Vitor Brito <joaovitor16611661@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.2"
django-rest-framework = "^0.1.0"
django-extensions = "3.2.3"
pytest = "^8.3.5"
factory-boy = "^3.3.3"
psycopg2-binary = "^2.9.10"
whitenoise = "5.3.0"
gunicorn = "^23.0.0"
gitpython = "^3.1.44"
django-cors-headers = "^3.13.0"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"